# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  imageUrl: string;
  additionalImages: string[];
  /** @nullable */
  badge?: string | null;
  isFeatured: boolean;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  /** @minLength 1 */
  imageUrl: string;
  additionalImages?: string[];
  /** @nullable */
  badge?: string | null;
  isFeatured?: boolean;
  isActive?: boolean;
  sortOrder?: number;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  category?: string;
  imageUrl?: string;
  additionalImages?: string[];
  /** @nullable */
  badge?: string | null;
  isFeatured?: boolean;
  isActive?: boolean;
  sortOrder?: number;
}

export type WorkshopEnquiryStatus = typeof WorkshopEnquiryStatus[keyof typeof WorkshopEnquiryStatus];


export const WorkshopEnquiryStatus = {
  pending: 'pending',
  contacted: 'contacted',
  confirmed: 'confirmed',
  completed: 'completed',
} as const;

export interface WorkshopEnquiry {
  id: number;
  name: string;
  phone: string;
  email: string;
  workshopType: string;
  peopleCount: number;
  preferredDate: string;
  preferredTime: string;
  message: string;
  status: WorkshopEnquiryStatus;
  createdAt: string;
}

export interface WorkshopEnquiryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 7 */
  phone: string;
  email: string;
  workshopType: string;
  /** @minimum 1 */
  peopleCount: number;
  preferredDate: string;
  preferredTime: string;
  message?: string;
}

export type WorkshopEnquiryUpdateStatus = typeof WorkshopEnquiryUpdateStatus[keyof typeof WorkshopEnquiryUpdateStatus];


export const WorkshopEnquiryUpdateStatus = {
  pending: 'pending',
  contacted: 'contacted',
  confirmed: 'confirmed',
  completed: 'completed',
} as const;

export interface WorkshopEnquiryUpdate {
  status: WorkshopEnquiryUpdateStatus;
}

export interface AdminLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  email?: string | null;
}

export interface AdminSummary {
  totalProducts: number;
  activeProducts: number;
  featuredProducts: number;
  workshopEnquiries: number;
}

export type ListProductsParams = {
category?: string;
featured?: boolean;
};