# lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSummary {
  totalProducts: number;
  activeProducts: number;
  featuredProducts: number;
  workshopEnquiries: number;
}