# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  category?: string;
  imageUrl?: string;
  additionalImages?: string[];
  /** @nullable */
  badge?: string | null;
  isFeatured?: boolean;
  isActive?: boolean;
  sortOrder?: number;
}