# lib/api-zod/src/generated/types/workshopEnquiry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkshopEnquiryStatus } from './workshopEnquiryStatus';

export interface WorkshopEnquiry {
  id: number;
  name: string;
  phone: string;
  email: string;
  workshopType: string;
  peopleCount: number;
  preferredDate: string;
  preferredTime: string;
  message: string;
  status: WorkshopEnquiryStatus;
  createdAt: Date;
}