# lib/api-zod/src/generated/types/workshopEnquiryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkshopEnquiryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 7 */
  phone: string;
  email: string;
  workshopType: string;
  /** @minimum 1 */
  peopleCount: number;
  preferredDate: string;
  preferredTime: string;
  message?: string;
}