# lib/api-zod/src/generated/types/workshopEnquiryUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkshopEnquiryUpdateStatus } from './workshopEnquiryUpdateStatus';

export interface WorkshopEnquiryUpdate {
  status: WorkshopEnquiryUpdateStatus;
}